Use int instead of size_t for vector indices

+++ include/vector.h
@@ -26,8 +26,8 @@
 
 #define vector_register(type) \
 struct vector_##type {\
-    size_t size;\
-    size_t capacity;\
+    int size;\
+    int capacity;\
     type *array;\
     type dummy;\
 };
@@ -44,7 +44,7 @@ struct vector_##type {\
 
 #define vector_push(vec, data) do {\
     if ((vec)->size == (vec)->capacity) {\
-        size_t capacity = (vec)->capacity ? 2*(vec)->capacity : 1;\
+        int capacity = (vec)->capacity ? 2*(vec)->capacity : 1;\
         void *array = realloc((vec)->array, capacity * sizeof((vec)->dummy));\
         if (array) {\
             (vec)->array = array;\
@@ -59,13 +59,13 @@ struct vector_##type {\
 #define vector_get(vec, i) ((vec)->array[i])
 
 #define vector_set(vec, j, data) do {\
-    if ((size_t)(j) < (vec)->capacity) {\
-        (vec)->array[(size_t)(j)] = (data);\
+    if ((j) < (vec)->capacity) {\
+        (vec)->array[(j)] = (data);\
     }\
 } while (0)
 
 #define vector_get_pointer(vec, i) \
-    (((size_t)(i) >= (vec)->capacity) ? NULL : (vec)->array + (size_t)(i))
+    ((i) >= (vec)->capacity) ? NULL : (vec)->array + (i)
 
 #define vector_destroy(vector) do {\
     free((vector)->array);\
@@ -73,5 +73,3 @@ struct vector_##type {\
     (vector)->size = 0;\
     (vector)->capacity = 0;\
 } while (0)
-
-
